working on the leveling system

+++ Loot.java
@@ -7,7 +7,7 @@ public class Loot {
     //WIP
     private ArrayList<Item> contents;
 
-    public Loot(int difficulty){
+    public Loot(double difficulty){
         contents = new ArrayList<Item>();
         for(int i = 0; i < (6 - (int)(difficulty * Math.random())); i++){
             contents.add(FileHandler.createItem("itm.dat"));
